Step 12.3: MongoDB Models created/defined
https://mongoosejs.com/docs/models.html

Models are fancy constructors compiled from Schema definitions. An instance of a model is called a document. Models are responsible for creating and reading documents from the underlying MongoDB database.

When we call mongoose.model() on a schema, Mongoose compiles a model for you. The first argument is the singular name of the collection your model is for. Mongoose automatically looks for the plural, lowercased version of your model name.

Example:  mongoose.model('User', UserSchema); Thus, for the example, the model "User" is for the "users" collection in the database.